+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.1.0-CR2
 
+* Bug fix related to self-node identification in the --peerNodes list when using FQDNs.
+
 ## 1.1.0-CR1
 
 * Roll back to Lucene to version 6.0.0 due to an issue in 6.2.1. See: